+++ CHANGELOG.md
@@ -61,7 +61,7 @@
 * Ikari III - The Rescue (World, 8-Way Joystick)
 * In Your Face
 * Joe & Mac Return
-* Ketsui
+* Ketsui Kizuna Jigoku Tachi
 * Knuckle Bash 2
 * Lethal Enforcers
 * Magical Drop
@@ -296,6 +296,7 @@
 * Backported some changes from later MAME for the ARM CPU core required for all the 32-bit Data East games to function
 correctly in this core affected drivers are backfire.c, deco32.c, deco156.c, deco_mlc.c and simpl1156.c
 * Fixed the Bally / Midway MCR games making them all playable in MAME2003-Plus [grant2258]
+* Fixed background character movements in Power Instinct 2 and Power Instinct Legends
 * Experimental support for MAME second-generation "Backdrop" artwork from pre-MAME 0.107. Backdrops are toggled via a core option and loaded from /libretro system/mame2003-plus/artwork. Right now there is a color when backdrops are first loaded.
 
 ------------------------------------